# apps/client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Olympus Protocol Metrics API",
    "version": "2.0.0",
    "description": "Cacheable REST API for Olympus protocol metrics, treasury assets, and OHM supply. /operations routes are deprecated compatibility routes."
  },
  "paths": {
...
    "/v2/treasury-assets/daily": {
      "get": {
        "summary": "Daily treasury assets",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC calendar start date."
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC calendar end date. Defaults to the latest published date."
          }
        ],
        "responses": {
          "200": {
            "description": "Treasury assets.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid date range.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v2/ohm-supply/daily": {
      "get": {
        "summary": "Daily OHM supply",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC calendar start date."
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC calendar end date. Defaults to the latest published date."
          }
        ],
        "responses": {
          "200": {
            "description": "OHM supply records.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid date range.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/operations/latest/tokenRecords": {
      "get": {
        "deprecated": true,
        "summary": "Latest treasury assets using the legacy tokenRecords shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/earliest/tokenRecords": {
      "get": {
        "deprecated": true,
        "summary": "Earliest treasury assets using the legacy tokenRecords shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/paginated/tokenRecords": {
      "get": {
        "deprecated": true,
        "summary": "Treasury assets using the legacy tokenRecords shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/latest/tokenSupplies": {
      "get": {
        "deprecated": true,
        "summary": "Latest OHM supply using the legacy tokenSupplies shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/earliest/tokenSupplies": {
      "get": {
        "deprecated": true,
        "summary": "Earliest OHM supply using the legacy tokenSupplies shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/paginated/tokenSupplies": {
      "get": {
        "deprecated": true,
        "summary": "OHM supply using the legacy tokenSupplies shape.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy WunderGraph-compatible response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/operations/atBlock/tokenRecords": {
      "get": {
        "deprecated": true,
        "summary": "Deprecated atBlock tokenRecords placeholder.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "501": {
            "description": "atBlock queries are not supported by the artifact-backed API.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/operations/atBlock/tokenSupplies": {
      "get": {
        "deprecated": true,
        "summary": "Deprecated atBlock tokenSupplies placeholder.",
        "parameters": [
          {
            "name": "wg_variables",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON-encoded legacy WunderGraph variables. dateOffset is accepted and ignored."
          }
        ],
        "responses": {
          "501": {
            "description": "atBlock queries are not supported by the artifact-backed API.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  }
}